+++ looper/project.py
@@ -20,7 +20,6 @@ from peppy.utils import make_abs_via_cfg
 from pipestat import PipestatError, PipestatManager
 from ubiquerg import expandpath, is_command_callable
 
-from .const import *
 from .exceptions import *
 from .pipeline_interface import PipelineInterface
 from .processed_project import populate_project_paths, populate_sample_paths
@@ -97,9 +96,23 @@ class Project(peppyProject):
         compute settings.
     """
 
-    def __init__(self, cfg, amendments=None, divcfg_path=None, runp=False, **kwargs):
+    def __init__(
+        self, cfg=None, amendments=None, divcfg_path=None, runp=False, **kwargs
+    ):
         super(Project, self).__init__(cfg=cfg, amendments=amendments)
+        prj_dict = kwargs.get("project_dict")
+
+        # init project from pephub:
+        if prj_dict is not None and cfg is None:
+            self.from_dict(prj_dict)
+            self["_config_file"] = os.getcwd()
+
         setattr(self, EXTRA_KEY, dict())
+
+        # add sample pipeline interface to the project
+        if kwargs.get(SAMPLE_PL_ARG):
+            self.set_sample_piface(kwargs.get(SAMPLE_PL_ARG))
+
         for attr_name in CLI_PROJ_ATTRS:
             if attr_name in kwargs:
                 setattr(self[EXTRA_KEY], attr_name, kwargs[attr_name])
@@ -688,6 +701,18 @@ class Project(peppyProject):
         for msg in msgs:
             _LOGGER.warning(msg)
         return samples_by_piface
+
+    def set_sample_piface(self, sample_piface: Union[List[str], str]) -> NoReturn:
+        """
+        Add sample pipeline interfaces variable to object
+
+        :param list | str sample_piface: sample pipeline interface
+        """
+        self._config.setdefault("sample_modifiers", {})
+        self._config["sample_modifiers"].setdefault("append", {})
+        self.config["sample_modifiers"]["append"]["pipeline_interfaces"] = sample_piface
+
+        self.modify_samples()
 
 
 def fetch_samples(